+++ basic/DrawableRect.hpp
@@ -6,7 +6,7 @@
 
 class DrawableRect : public IDrawable {
 public:
-    DrawableRect(int width, int height);
+    DrawableRect(float width, float height);
 
     virtual void draw(IObject *model, sf::RenderTarget *drawer);
